Refactoring with Go project layout
