+++ docs/images/Movement_between_registries.png
