+++ docs/images/Notary_client_targets_proposal.png
