+++ docs/images/Snapshot_merkle_tree.png
