+++ test/tuf-repo/staged/root.json
